Post prepayment invoices to a single endpoint
The form decides whether to generate or schedule from the send date.
