+++ corehq/apps/domain/templates/domain/bootstrap3/current_subscription.html
@@ -13,7 +13,6 @@
   {% initial_page_data "can_change_subscription" can_change_subscription %}
   {% registerurl "domain_credits_payment" domain %}
   {% registerurl "domain_wire_payment" domain %}
-  {% registerurl "domain_schedule_prepayment_invoice" domain %}
 
   <div class="row">
     <div class="col-sm-12">
