+++ corehq/apps/domain/urls.py
@@ -32,7 +32,6 @@ from corehq.apps.domain.views.accounting import (
     GeneralPlanQuestionView,
     InternalSubscriptionManagementView,
     InvoiceStripePaymentView,
-    SchedulePrepaymentInvoiceView,
     SelectedAnnualPlanView,
     SelectedCustomPlanView,
     SelectedEnterprisePlanView,
@@ -168,9 +167,6 @@ domain_settings = [
         name=CreditsStripePaymentView.urlname),
     url(r'^subscription/credits/make_wire_payment/$', CreditsWireInvoiceView.as_view(),
         name=CreditsWireInvoiceView.urlname),
-    url(r'^subscription/credits/schedule_prepayment_invoice/$',
-        SchedulePrepaymentInvoiceView.as_view(),
-        name=SchedulePrepaymentInvoiceView.urlname),
     url(r'^billing/statements/download/(?P<statement_id>[\w-]+).pdf$',
         BillingStatementPdfView.as_view(),
         name=BillingStatementPdfView.urlname),
