+++ corehq/util/test_utils.py
@@ -235,6 +235,7 @@ class privilege_enabled:
         'corehq.apps.data_dictionary.views.domain_has_privilege',
         'corehq.apps.data_cleaning.utils.cases.domain_has_privilege',
         'corehq.apps.domain.auth.domain_has_privilege',
+        'corehq.apps.domain.views.accounting.has_privilege',
         'corehq.apps.export.views.list.domain_has_privilege',
         'corehq.apps.export.views.new.domain_has_privilege',
         'corehq.apps.users.landing_pages.domain_has_privilege',
